requirements.txt
Flask
python-dotenv
openai
requests
numpy
collections